Adding smartparens config, using paredit bindings
There is some issue where "C-M-d" with mac-command-key-is-meta, emacs
refuses to see that key combo entered (it doesn't show up in the help
window even). Probably somthing is intercepting the call but turning on
option as another meta key for now since that seems to work. Still want
to figure that one out since the option key as meta is still inconvenient